CFDomains

Documentation for CFDomains.

CFDomains.AbstractDomain

CFDomains.FDDomain

CFDomains.HVLayout

CFDomains.HyperDiffusion

CFDomains.PressureCoordinate

CFDomains.Shell

CFDomains.SigmaCoordinate

CFDomains.SpectralDomain

CFDomains.SpectralSphere

CFDomains.VHLayout

CFDomains.VerticalCoordinate

CFDomains.allocate_field

CFDomains.allocate_fields

CFDomains.data_layout

CFDomains.filter!

CFDomains.laplace_dx

CFDomains.mass_level

CFDomains.periodize!

CFDomains.pressure_level

CFDomains.scratch_space

CFDomains.shell

CFDomains.AbstractDomain — Type

Parent type of SpectralDomain and FDDomain

CFDomains.FDDomain — Type

FDDomain <: AbstractDomain`

CFDomains.HVLayout — Type

struct HVLayout{rank} end
layout = HVLayout(rank)

Singleton type describing a multi-layer data layout where horizontal layers are contiguous in memory. rank is the number of horizontal indices (1 or 2).

CFDomains.HyperDiffusion — Method

filter = Hyperdiffusion(domain, niter, nu, fieldtype::Symbol)

Return a filter that applies Laplacian diffusion iterated niter times with hyperdiffusive coefficient nu ≥ 0 on fields of type fieldtype. Supported field types for domain::VoronoiSphere are :scalar, :vector_curl, :vector_div.

CFDomains.PressureCoordinate — Type

abstract type PressureCoordinate{N} <: VerticalCoordinate{N} end

Parent type for a pressure-based vertical coordinate. Children types should specialize pressure_level and mass_level. See also VerticalCoordinate.

CFDomains.Shell — Type

multi_layer_domain = Shell(nz::Int, layer::AbstractDomain, layout)

Return a multi-layer domain made of nz layers with data layout specified by layout. Unless you know what you are doing, it is recommended to use rather:

multi_layer_domain = shell(nz::Int, layer::AbstractDomain)

which gets the data layout from data_layout(layer). Otherwise, multi_layer_domain may be non-optimal or non-usable.

CFDomains.SigmaCoordinate — Type

sigma = SigmaCoordinate(N, ptop) <: PressureCoordinate{N}

Pressure based sigma-coordinate for N levels with top pressure ptop. Pressure levels are linear in vertical coordinate k : k/N = (ps-p)/(ps-ptop) where k ranges from 0 (ground) to N (model top).

CFDomains.SpectralDomain — Type

SpectralDomain <: AbstractDomain

parent type of SpectralSphere

CFDomains.SpectralSphere — Type

Parent type for spherical domains using spherical harmonics.

CFDomains.VHLayout — Type

struct VHLayout{rank} end
layout = VHLayout(rank)

Singleton type describing a multi-layer data layout where vertical columnsare contiguous. rank is the number of horizontal indices (1 or 2).

CFDomains.VerticalCoordinate — Type

abstract type VerticalCoordinate{N} end

Parent type for generalized vertical coordinates ranging from 0 to N. See also PressureCoordinate.

CFDomains.allocate_field — Function

field = allocate_field(kind::Symbol, domain::AbstractDomain, precision::Type)

Allocate a field of the given kind and precision over the given domain. Typical values for precision are Float32, Float64 or ForwardDiff.Dual. Depending on the domain, valid values for kind may include :scalar, :vector, :scalar_spec, scalar_spat (spectral/spatial representation of a scalar field), :vector_spec, vector_spat (spectral/spatial representation of a vector field).

Internally, allocate_field(kind::Symbol, domain, F) returns allocate_field(Val(kind), domain, F). To specialize allocate_field for MyDomain <: AbstractDomain, one must provide methods for:

allocate_field(::Val{kind}, domain::MyDomain, F)

where symbol kind::Symbol is one of the valid field kinds for that domain.

CFDomains.allocate_fields — Function

fields = allocate_fields(kinds::Tuple, domain::AbstractDomain, F::Type)
fields = allocate_fields(kinds::NamedTuple, domain::AbstractDomain, F::Type)

Allocate a (named) tuple of fields according to the provided kinds. For instance:

fields = allocate_fields((:vector, :scalar), domain, F)
fields = allocate_fields((a=:vector, b=:scalar), domain, F)

are equivalent to, respectively:

fields = (allocate_field(:vector, domain, F), allocate_field(:scalar, domain, F))
fields = (a=allocate_field(:vector, domain, F), b=allocate_field(:scalar, domain, F))

CFDomains.data_layout — Method

layout = data_layout(shell::Shell)

Return layout describing the data layout of multi-layer domain shell.

layout = data_layout(domain::Domain)

Return layout describing the preferred data layout for a shell made of layers of type Domain.

Typical values for layout are the singletons HVLayout() (layers are contiguous in memory) and VHLayout (columns are contiguous in memory).

CFDomains.filter! — Function

# generic
filtered = filter!(space, scratch, field, filter::AbstractFilter, dt, mgr::LoopManager)

# out-of-place, allocates `scratch`, allocates and returns `space`
filtered = filter!(void, void, field, filter, dt, mgr)

# mutating, non-allocating
filter!(filtered, scratch, field, filter, dt, mgr)

# in-place, non-allocating
filter!(field, scratch, field, filter, dt, mgr)

Apply filter to field during time dt with loop manager mgr. mgr can be nothing. space is the output field, or void. In the latter case, space is allocated as similar(field) or equivalent. scratch is scratch space allocated with scratch_space, or void. In the latter case, scratch is implicitly allocated.

CFDomains.laplace_dx — Method

Estimates the largest eigenvalue -lambda=dx^-2 of the scalar Laplace operator and returns dx which is a (non-dimensional) length on the unit sphere characterizing the mesh resolution. By design, the Courant number for the wave equation with unit wave speed solved with time step dt is dt/dx.

CFDomains.mass_level — Function

m = mass_level(k, masstot, vcoord::PressureCoordinate{N})

Return mass m in level k/2 as prescribed by vertical coordinate vcoord and total mass masstot, i.e. surface pressure minus top pressure.

So-called full levels correspond to odd values k=1,2...2N-1 while interfaces between full levels (so-called half-levels) correspond to even values k=0,2...2N

CFDomains.periodize! — Method

periodize!(data, box::AbstractBox, mgr)

Enforce horizontally-periodic boundary conditions on array data representing grid point values in box. data may also be a collection, in which case periodize! is applied to each element of the collection. Call periodize! on data obtained by computations involving horizontal averaging/differencing.

CFDomains.pressure_level — Function

p = pressure_level(k, ps, vcoord::PressureCoordinate{N})

Returns pressure p corresponding to level k/2 as prescribed by vertical coordinate vcoord and surface pressure ps.

So-called full levels correspond to odd values k=1,2...2N-1 while interfaces between full levels (so-called half-levels) correspond to even values k=0,2...2N

CFDomains.scratch_space — Function

scratch = scratch_space(filter::AbstractFilter, field, [scratch])

If scratch is omitted or ::Void, allocate and return scratch space for applying filter to field. Otherwise just return scratch. See also filter!.

CFDomains.shell — Method

multi_layer_domain = shell(nz::Int, layer::AbstractDomain)

Return a multi-layer domain made of nz layers with data layout specified by data_layout(layer).